refactor: move di classes to core module
